FitMind/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.58.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
slowapi==0.1.9
